requirements.txt
streamlit
openai
python-dotenv
streamlit_ace